% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schrute.R
\docType{package}
\name{schrute}
\alias{schrute}
\alias{schrute-package}
\title{schrute: The Entire Transcript from The Office in Tidy Format.}
\description{
Analyze and have fun with the text from the best series of all time
}
