% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-theoffice.R
\docType{data}
\name{theoffice}
\alias{theoffice}
\title{The entire script transcriptions from The Office}
\format{A tibble with 55130 observations of 9 variables:
\describe{
\item{index}{row index}
\item{season}{season number}
\item{episode}{episode number}
\item{episode_name}{name of the episode}
\item{director}{director of the episode}
\item{writer}{name(s) of writers of the episode}
\item{character}{name of the character saying the line}
\item{text}{words spoken by that actor}
\item{text_w_direction}{words spoken by that actor with stage direction included}
}}
\source{
\url{https://transcripts.foreverdreaming.org}
}
\usage{
theoffice
}
\description{
The entire script transcriptions from The Office
}
\keyword{datasets}
